fix(build): move list-options.ts under src folder
